+++ src/modules/meal/meal.service.ts
@@ -7,12 +7,14 @@ import { QueryResult } from '../../common/interfaces';
 import { CreateMealDto } from './meal.dto';
 import { BadRequestException } from '../../exceptions/bad-request.exception';
 import { NotFoundException } from '../../exceptions/not-found.exception';
+import { LoggerService } from '../logger/logger.service';
 
 @Injectable()
 export class MealService {
 
     constructor(@InjectModel(models.MEAL_MODEL)
-                private mealModel: Model<MealDocument>) {}
+                private mealModel: Model<MealDocument>,
+                private loggerService: LoggerService) {}
 
     async create(createMealDto: CreateMealDto): Promise<QueryResult<MealDocument>> {
         const createdMeal = new this.mealModel(createMealDto);
@@ -35,26 +37,26 @@ export class MealService {
     }
 
     async find(id: string): Promise<QueryResult<MealDocument>> {
-        const  context = 'MealService/find:';
+        const context = 'MealService/find';
 
         if (!Types.ObjectId.isValid(id)) {
             const message = `Provided "${id}" that is not a correct MongoDB id.`;
-            console.error(context, message);
+            this.loggerService.error(context, message);
 
             throw new BadRequestException(context, message);
         }
 
         const meal = await this.mealModel.findById(id);
 
-        if (meal === null) {
+        if (!meal) {
             const message = `Cannot find a meal with "${id}" id.`;
-            console.error(context, message);
+            this.loggerService.error(context, message);
 
             throw new NotFoundException(context, message);
         }
 
         const message = `Found meal with "${id}" id.`;
-        console.info(context, message);
+        this.loggerService.info(context, message);
 
         return {
             data: meal as MealDocument,